+++ iceoryx_posh/include/iceoryx_posh/popo/publisher.hpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2020 by Robert Bosch GmbH. All rights reserved.
-// Copyright (c) 2020 - 2021 by Apex.AI Inc. All rights reserved.
+// Copyright (c) 2020 - 2022 by Apex.AI Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,95 +18,23 @@
 #ifndef IOX_POSH_POPO_TYPED_PUBLISHER_HPP
 #define IOX_POSH_POPO_TYPED_PUBLISHER_HPP
 
-#include "iceoryx_hoofs/cxx/type_traits.hpp"
-#include "iceoryx_posh/internal/popo/typed_port_api_trait.hpp"
-#include "iceoryx_posh/popo/base_publisher.hpp"
-#include "iceoryx_posh/popo/sample.hpp"
+#include "iceoryx_posh/internal/popo/publisher_impl.hpp"
 
 namespace iox
 {
 namespace popo
 {
-///
-/// @brief The PublisherInterface class defines the publisher interface used by the Sample class to make it generic.
-/// This allows any publisher specialization to be stored as a reference by the Sample class.
-/// It is also needed to avoid circular dependencies between Sample and Publisher.
-///
-template <typename T, typename H>
-class PublisherInterface
-{
-  public:
-    using SampleType = Sample<T, H>;
-
-    virtual void publish(SampleType&& sample) noexcept = 0;
-
-  protected:
-    PublisherInterface() = default;
-};
-template <typename T, typename H = mepoo::NoUserHeader, typename BasePublisher_t = BasePublisher<>>
-class PublisherImpl : public BasePublisher_t, private PublisherInterface<T, H>
-{
-    using DataTypeAssert = typename TypedPortApiTrait<T>::Assert;
-    using HeaderTypeAssert = typename TypedPortApiTrait<H>::Assert;
-
-  public:
-    PublisherImpl(const capro::ServiceDescription& service,
-                  const PublisherOptions& publisherOptions = PublisherOptions());
-    PublisherImpl(const PublisherImpl& other) = delete;
-    PublisherImpl& operator=(const PublisherImpl&) = delete;
-    PublisherImpl(PublisherImpl&& rhs) = default;
-    PublisherImpl& operator=(PublisherImpl&& rhs) = default;
-    virtual ~PublisherImpl() = default;
-
-    ///
-    /// @brief loan Get a sample from loaned shared memory and consctruct the data with the given arguments.
-    /// @param args Arguments used to construct the data.
-    /// @return An instance of the sample that resides in shared memory or an error if unable ot allocate memory to
-    /// loan.
-    /// @details The loaned sample is automatically released when it goes out of scope.
-    ///
-    template <typename... Args>
-    cxx::expected<Sample<T, H>, AllocationError> loan(Args&&... args) noexcept;
-
-    ///
-    /// @brief publish Publishes the given sample and then releases its loan.
-    /// @param sample The sample to publish.
-    ///
-    void publish(Sample<T, H>&& sample) noexcept override;
-
-    ///
-    /// @brief publishCopyOf Copy the provided value into a loaned shared memory chunk and publish it.
-    /// @param val Value to copy.
-    /// @return Error if unable to allocate memory to loan.
-    ///
-    cxx::expected<AllocationError> publishCopyOf(const T& val) noexcept;
-    ///
-    /// @brief publishResultOf Loan a sample from memory, execute the provided callable to write to it, then publish it.
-    /// @param c Callable with the signature void(T*, ArgTypes...) that write's it's result to T*.
-    /// @param args The arguments of the callable.
-    /// @return Error if unable to allocate memory to loan.
-    ///
-    template <typename Callable, typename... ArgTypes>
-    cxx::expected<AllocationError> publishResultOf(Callable c, ArgTypes... args) noexcept;
-
-  protected:
-    using BasePublisher_t::port;
-
-  private:
-    Sample<T, H> convertChunkHeaderToSample(mepoo::ChunkHeader* const header) noexcept;
-
-    cxx::expected<Sample<T, H>, AllocationError> loanSample() noexcept;
-
-    using PublisherSampleDeleter = SampleDeleter<typename BasePublisher_t::PortType>;
-    PublisherSampleDeleter m_sampleDeleter{port()};
-};
-
+/// @brief The Publisher class for the publish-subscribe messaging pattern in iceoryx.
+/// @param[in] T user payload type
+/// @param[in] H user header type
 template <typename T, typename H = mepoo::NoUserHeader>
-using Publisher = PublisherImpl<T, H>;
+class Publisher : public PublisherImpl<T, H>
+{
+  public:
+    using PublisherImpl<T, H>::PublisherImpl;
+};
 
 } // namespace popo
 } // namespace iox
-
-#include "iceoryx_posh/internal/popo/publisher.inl"
 
 #endif // IOX_POSH_POPO_TYPED_PUBLISHER_HPP
